requirements.txt
pandas
matplotlib
python-telegram-bot